apex: fix task_pid and thread_id calculation
If the physical address of a task or pid happened to have the top bit
set (common on nommu systems) these functions would return stupid values
as they were performing signed shifts rather than unsigned.

+++ sys/kern/thread.c
@@ -281,7 +281,7 @@ int
 thread_id(struct thread *t)
 {
 	const unsigned shift = floor_log2(alignof(struct thread));
-	return (int)virt_to_phys(t) >> shift;
+	return (unsigned long)virt_to_phys(t) >> shift;
 }
 
 /*
